Impliment several interfaces
