+++ src/main/java/miniprogram/server/mapper/WxMpBindInfoMapper.java
@@ -19,4 +19,5 @@ public interface WxMpBindInfoMapper{
 
     WxMpBindInfo select(@RequestBody WxMpBindInfo wxMpBindInfo);
 
+    void insert(WxMpBindInfo wxMpBindInfo);
 }
