+++ src/main/resources/mapper/WxMpBindInfoMapper.xml
@@ -16,7 +16,7 @@
   </resultMap>
 
   <select id="selectByOrgId" resultType="miniprogram.server.beans.Organization">
-    select o.id, o.corpname, o.corp_code from wx_mp_bind_info w left join organization o on w.org_id = o.id where w.wx_uid = #{wxUid} and w.isbind = 1
+    select o.id, o.corpname, o.corp_code from wx_mp_bind_info w left join organization o on w.org_id = o.corp_code where w.wx_uid = #{wxUid} and w.isbind = 1
   </select>
 
   <select id="selectByOrgIdAndUid" resultMap="BaseResultMap">
@@ -49,9 +49,9 @@
 
   <insert id="insert" parameterType="miniprogram.server.beans.WxMpBindInfo">
     insert into wx_mp_bind_info(org_id, wx_uid, username, isbind)
-    values (#{wxUid, jdbcType=VARCHAR},
-            #{orgId, jdbcType=VARCHAR},
-            #{userid, jdbcType=INTEGER},
+    values (#{orgId, jdbcType=INTEGER},
+            #{wxUid, jdbcType=INTEGER},
+            #{username, jdbcType=VARCHAR},
             #{isbind, jdbcType=INTEGER})
   </insert>
